Deliver endpoints to the accept queue synchronously when possible
Before this change, when a new connection was created after receiving
an ACK that matched a SYN-cookie, it was always delivered asynchronously
to the accept queue. There was a chance that the listening endpoint
would process a SYN from another client before the delivery happen, and
the listening endpoint would not know yet that the queue was about to be
full, once the delivery happened.

Now, whenever possible, the delivery of connections matching a SYN
cookie happens synchronously. And when it is not possible, it is done in
a goroutine but this connection counts as if it was in the accept queue
already, when checking if the backlog is full or not.

Fixes #6545

PiperOrigin-RevId: 394742288